fix:finished

+++ inventory_management/modules/product/models.py
@@ -15,7 +15,7 @@ class Product(BaseModel):
     id = models.CharField(primary_key=True, unique=True, max_length=50, )
     unit = models.CharField(max_length=20, default="N/A")
     price = models.DecimalField(max_digits=10, decimal_places=2, default=0)
-    amount = models.BigIntegerField(default=0)
+    amount = models.DecimalField(max_digits=10, decimal_places=2, default=0)
 
     store = models.ForeignKey(
         Store,
